app/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arize-phoenix-otel",
    "deep-sort-realtime>=1.3.2",
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "langchain>=1.2.10",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "minio>=7.2.20",
    "openinference-instrumentation-langchain",
    "opencv-python>=4.13.0.92",
    "ovmsclient>=2023.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "setuptools>=68.0.0,<82",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tritonclient[grpc]>=2.42.0",
]
